requirements.txt
asgiref==3.2.7
Django==3.0.6
pytz==2020.1
sqlparse==0.3.1
asgiref==3.2.7
Django==3.0.6
django-filter==2.2.0
djangorestframework==3.11.0
importlib-metadata==1.6.0
Markdown==3.2.2
pytz==2020.1
sqlparse==0.3.1
zipp==3.1.0
asgiref==3.2.7
Django==3.0.6
django-filter==2.2.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
importlib-metadata==1.6.0
Markdown==3.2.2
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1
zipp==3.1.0
